560 renaming housing recovery to housing valuation recovery (#565)
* housing valuation recovery new classes

* the old method call the new one

* change log and sphinx docs

* add deprecation tag to sphinx

* add deprecated library

* update more requirements

* add copyright year
