+++ pyincore/analyses/housingvaluationrecovery/__init__.py
@@ -1,0 +1,8 @@
+# Copyright (c) 2019 University of Illinois and others. All rights reserved.
+#
+# This program and the accompanying materials are made available under the
+# terms of the Mozilla Public License v2.0 which accompanies this distribution,
+# and is available at https://www.mozilla.org/en-US/MPL/2.0/
+
+
+from pyincore.analyses.housingvaluationrecovery.housingvaluationrecovery import HousingValuationRecovery
